+++ include/udev.h
@@ -3,7 +3,10 @@
 
 #include "main.h"
 
+// checks if the device has a tty device associated with it
 int has_tty(struct udevInfo udev);
+
+// initalizer functions
 int init_udev(struct udevInfo* udev);
 void deinit_udev(struct udevInfo* udev);