+++ src/main.c
@@ -1,14 +1,6 @@
 /*
  * This is a PoC that implements some ideas which should help protect against
  * rubber ducky attacks as a daemon
- *
- * TODO:
- * 		* implement microcontroller firmware overwrite over the serial port
- * 		Maybe?
- *      * replace strcmp_ss with safelib implementation?
- * 		* write unit tests with cmocka?
- *
- * TODO BUGS:
  * */
 
 #include <stdbool.h>
@@ -48,6 +40,10 @@
 #define T struct deviceInfo
 #include "mbuffertemplate.h"
 
+#define PREFIX struct_timespec
+#define T struct timespec
+#include "mbuffertemplate.h"
+
 // global variables
 bool brexit;
 bool reloadconfig;
@@ -75,26 +71,29 @@ static int deinit_device(struct deviceInfo* device, struct configInfo* config, s
         device->fd = -1;
     }
 
-    if (config->logkeys && device->score < 0) {
+	// if keylogging was enabled
+    if (config->logkeys ) {
         if (device->devlog.b != NULL) {
-            if (device->devlog.size > 0) {
+            if (device->score < 0) {
                 if (m_append_array_char(&device->devlog, "\n\0", 2)) {
                     LOG(-1, "append_mbuffer_array_char failed!\n");
                 }
                 LOG(1, "-> %s\n", device->devlog.b);
 
+				// write keylog to the log file
                 if (write(kbd->outfd, (char*)device->devlog.b, device->devlog.size) < 0) {
                     ERR("write");
                 }
-
-                m_free(&device->devlog);
             }
+            m_free(&device->devlog);
         }
 
         if (config->xkeymaps && device->xstate != NULL) {
             xkb_state_unref(device->xstate);
             device->xstate = NULL;
         }
+
+        m_free(&device->strokesdiff);
     }
     return err;
 
@@ -132,14 +131,16 @@ static int remove_fd(struct managedBuffer* device, struct configInfo* config, st
         {
             size_t i;
             size_t bigsize = 0;
-
-            for (i = 0; i < device->size; i++) { // find the biggest fd in the array
+			
+			// find the biggest fd in the array
+            for (i = 0; i < device->size; i++) { 
                 if (m_deviceInfo(device)[i].openfd[0] != '\0') {
                     bigsize = i + 1;
                 }
             }
-
-            if (bigsize < device->size) { // free the unnecessary space
+			
+			// free the unnecessary space
+            if (bigsize < device->size) { 
                 bool failed = false;
 
                 for (i = bigsize; i < device->size; i++) {
@@ -158,7 +159,7 @@ static int remove_fd(struct managedBuffer* device, struct configInfo* config, st
         }
 
     } else {
-        LOG(-1, "Did not find %d\n", fd);
+        LOG(1, "Did not find %d\n", fd);
         return -2;
     }
 
@@ -173,15 +174,17 @@ static int add_fd(struct managedBuffer* device, struct keyboardInfo* kbd, struct
     int fd;
 
     LOG(1, "Adding: %s\n", location);
-
-    fd = open(location, O_RDONLY | O_NONBLOCK); // open a fd to the devnode
+	
+	// open a fd to the devnode
+    fd = open(location, O_RDONLY | O_NONBLOCK); 
     if (fd < 0) {
         LOG(-1, "open failed\n");
         err = -1;
         goto error_exit;
     }
 
-    if (device->size <= fd) { // allocate more space if the fd doesn't fit
+	// allocate more space if the fd doesn't fit
+    if (device->size <= fd) {
         size_t i;
         size_t prevsize;
 
@@ -192,35 +195,50 @@ static int add_fd(struct managedBuffer* device, struct keyboardInfo* kbd, struct
             goto error_exit;
         }
 
-        for (i = prevsize; i < device->size; i++) { // initialize all members of the device array which haven't been used
+		// initialize all members of the device array which haven't been used
+        for (i = prevsize; i < device->size; i++) { 
             m_deviceInfo(device)[i].openfd[0] = '\0';
             m_deviceInfo(device)[i].fd = -1;
 
             m_deviceInfo(device)[i].score = 0;
             m_deviceInfo(device)[i].xstate = NULL;
             m_init(&m_deviceInfo(device)[i].devlog, sizeof(char));
-            memset_s(&m_deviceInfo(device)[i].time_added, sizeof(struct timespec), 0);
+
+            m_init(&m_deviceInfo(device)[i].strokesdiff, sizeof(struct timespec));
+			m_deviceInfo(device)[i].currdiff = 0;
         }
     }
 
-    if (m_deviceInfo(device)[fd].openfd[0] == '\0' && m_deviceInfo(device)[fd].fd == -1) { // allocate and set the openfd
+	// allocate and set the openfd
+    if (m_deviceInfo(device)[fd].openfd[0] == '\0' && m_deviceInfo(device)[fd].fd == -1) {
         strcpy_s(m_deviceInfo(device)[fd].openfd, MAX_SIZE_PATH, location);
         m_deviceInfo(device)[fd].fd = fd;
 
-        if (clock_gettime(CLOCK_REALTIME, &m_deviceInfo(device)[fd].time_added)) {
-            ERR("clock_gettime");
-            m_deviceInfo(device)[fd].time_added.tv_sec = 0;
-            m_deviceInfo(device)[fd].time_added.tv_nsec = 0;
-        }
-
         if (config->logkeys && config->xkeymaps) {
+            size_t i;
+
+			// set the device state
             m_deviceInfo(device)[fd].xstate = xkb_x11_state_new_from_device(kbd->x.keymap, kbd->x.con, kbd->x.device_id);
             if (!m_deviceInfo(device)[fd].xstate) {
                 LOG(-1, "xkb_x11_state_new_from_device failed!\n");
-                return -1;
+				err = -3;
+				goto error_exit;
+            }
+
+			// allocate the array 
+            if(m_realloc(&m_deviceInfo(device)[fd].strokesdiff, 6)) {
+				LOG(-1, "m_realloc failed!\n");
+				err = -4;
+				goto error_exit;
+			}
+
+			// set every member to 0
+            for(i = 0; i < m_deviceInfo(device)[fd].strokesdiff.size; i++) {
+                memset_s(m_struct_timespec(&m_deviceInfo(device)[fd].strokesdiff), sizeof(struct timespec), 0);
             }
         }
 
+		// add a new fd which should be polled
         {
             struct epoll_event event;
             memset_s(&event, sizeof(struct epoll_event), 0);
@@ -235,7 +253,7 @@ static int add_fd(struct managedBuffer* device, struct keyboardInfo* kbd, struct
         }
 
     } else {
-        LOG(-1, "Somehow this element is already in use! This is a bug and should not happen!\n");
+        LOG(-1, "Somehow this element is already in use! There is something seriously wrong!\n");
         err = -5;
         goto error_exit;
     }
@@ -255,6 +273,7 @@ static int handle_udevev(struct managedBuffer* device, struct keyboardInfo* kbd,
     const char* devnode;
     const char* action;
 
+	// get a device context from the monitor
     udev->dev = udev_monitor_receive_device(udev->mon);
     if (udev->dev == NULL) {
         LOG(-1, "udev_monitor_receive_device failed\n");
@@ -264,15 +283,17 @@ static int handle_udevev(struct managedBuffer* device, struct keyboardInfo* kbd,
 
     devnode = udev_device_get_devnode(udev->dev);
     action = udev_device_get_action(udev->dev);
-
-    if (devnode != NULL && action != NULL) { // device has a devnode
-        LOG(1, "%s %s -> %s [%s] | %s:%s\n",
+	
+	// device has a devnode
+    if (devnode != NULL && action != NULL) { 
+        LOG(2, "%s %s -> %s [%s] | %s:%s\n",
             udev_device_get_devpath(udev->dev),
             action, devnode, udev_device_get_subsystem(udev->dev),
             udev_device_get_property_value(udev->dev, "MAJOR"),
             udev_device_get_property_value(udev->dev, "MINOR"));
-
-        if (strncmp_ss(action, "add", 3) == 0) { // add the devnode to the array
+		
+		// add the devnode to the array
+        if (strncmp_ss(action, "add", 3) == 0) {
             int fd;
 
             fd = add_fd(device, kbd, config, epollfd, devnode);
@@ -285,15 +306,17 @@ static int handle_udevev(struct managedBuffer* device, struct keyboardInfo* kbd,
                 err = -2;
                 goto error_exit;
             }
-
-        } else if (strncmp_ss(action, "remove", 6) == 0) { // remove it
+	
+		// remove the fd from the device array
+        } else if (strncmp_ss(action, "remove", 6) == 0) {
             int fd;
 
+			// search for the fd and remove it if possible
             fd = search_fd(device, devnode);
             if (fd >= 0) {
                 if (remove_fd(device, config, kbd, epollfd, fd)) {
                     LOG(-1, "remove_fd failed\n");
-                    err = -4;
+                    err = -3;
                     goto error_exit;
                 }
             }
@@ -303,44 +326,6 @@ static int handle_udevev(struct managedBuffer* device, struct keyboardInfo* kbd,
 error_exit:
     udev_device_unref(udev->dev);
     return err;
-}
-
-static int get_nexttime(struct managedBuffer* device, struct configInfo* config)
-{
-    if (device->size > 0) {
-        size_t i;
-        struct timespec smallest, curr, diff;
-
-        // calculate the biggest number that fits in the type time_t
-        // workaround for gcc using a 32 bit int for calculations when using the shift operator
-        smallest.tv_sec = 0x1;
-        smallest.tv_sec <<= (sizeof(time_t) * 8 - 1);
-        smallest.tv_sec = ~smallest.tv_sec;
-
-        smallest.tv_nsec = LONG_MAX;
-
-        for (i = 0; i < device->size; i++) {
-            struct timespec current = m_deviceInfo(device)[i].time_added;
-
-            if (current.tv_sec > 0 && current.tv_nsec > 0) {
-                if (smallest.tv_sec > current.tv_sec || (smallest.tv_sec == current.tv_sec && smallest.tv_nsec > current.tv_nsec)) {
-                    smallest.tv_nsec = current.tv_nsec;
-                    smallest.tv_sec = current.tv_sec;
-                }
-            }
-        }
-
-        if (clock_gettime(CLOCK_REALTIME, &curr)) {
-            ERR("clock_gettime");
-            return -1;
-        }
-
-        diff.tv_sec = config->maxtime.tv_sec - (curr.tv_sec - smallest.tv_sec);
-        diff.tv_nsec = config->maxtime.tv_nsec - (curr.tv_nsec - smallest.tv_nsec);
-
-        return diff.tv_sec * 1000 + round(diff.tv_nsec / 1000000);
-    }
-    return -1;
 }
 
 int main(int argc, char* argv[])
@@ -439,16 +424,14 @@ int main(int argc, char* argv[])
         ERR("epoll_ctl");
         return -1;
     }
+	LOG(0,"Startup done!\n");
 
     // MAIN LOOP
     while (!brexit) {
-        int readfds, timeout;
+        int readfds;
         struct epoll_event events[MAX_SIZE_EVENTS];
 
-        timeout = get_nexttime(&device, &config);
-        LOG(1, "Next timeout: %d\n", timeout);
-
-        readfds = epoll_wait(epollfd, events, MAX_SIZE_EVENTS, timeout);
+        readfds = epoll_wait(epollfd, events, MAX_SIZE_EVENTS, -1);
         if (readfds < 0) {
             if (errno == EINTR) { // fix endless loop when receiving SIGHUP
                 readfds = 0;
@@ -479,18 +462,10 @@ int main(int argc, char* argv[])
                     } else {
                         // handle keyboard grabbing
                         if (event.type == EV_KEY) {
-                            LOG(1, "fd=%d event.type=%d event.code=%d event.value=%d\n", fd, event.type, event.code, event.value);
-                            if (m_deviceInfo(&device)[fd].score >= config.maxcount) {
-                                size_t k;
-                                bool handle = false;
-
-                                for (k = 0; k < config.blacklist.size; k++) {
-                                    if (event.code == ((long int*)config.blacklist.b)[k]) {
-                                        handle = true;
-                                    }
-                                }
-
-                                if (handle && event.value == 0) {
+                            LOG(2, "fd=%d event.type=%d event.code=%d event.value=%d\n", fd, event.type, event.code, event.value);
+                            
+							if (m_deviceInfo(&device)[fd].score >= config.maxcount) {
+								if (event.value == 0) {
                                     int ioctlarg = 1;
                                     if (ioctl(fd, EVIOCGRAB, &ioctlarg)) {
                                         ERR("ioctl");
@@ -513,31 +488,7 @@ int main(int argc, char* argv[])
             }
             events[i].events = 0;
         }
-
-        // handle timeout
-        for (i = 0; i < device.size; i++) {
-            if (m_deviceInfo(&device)[i].time_added.tv_sec != 0 && m_deviceInfo(&device)[i].time_added.tv_nsec != 0) {
-                struct timespec diff, curr;
-                if (clock_gettime(CLOCK_REALTIME, &curr)) {
-                    ERR("clock_gettime");
-                    curr.tv_sec = 0;
-                    curr.tv_nsec = 0;
-                }
-
-                diff.tv_sec = curr.tv_sec - m_deviceInfo(&device)[i].time_added.tv_sec; // get time difference
-                diff.tv_nsec = curr.tv_nsec - m_deviceInfo(&device)[i].time_added.tv_nsec;
-
-                // check for timeout
-                if (diff.tv_sec >= config.maxtime.tv_sec && diff.tv_nsec >= config.maxtime.tv_nsec) {
-                    if (remove_fd(&device, &config, &kbd, epollfd, i)) {
-                        LOG(-1, "remove_fd failed\n");
-                    }
-                    LOG(0, "%d timed out > %ds %dns timestamp:%d %d\n", i, diff.tv_sec,
-                        diff.tv_nsec, curr.tv_sec, curr.tv_nsec);
-                }
-            }
-        }
-
+        
         // reload config if SIGHUP is received
         if (reloadconfig) {
             LOG(0, "Reloading config file...\n");
@@ -567,8 +518,6 @@ int main(int argc, char* argv[])
         ERR("close");
         return -1;
     }
-
-    m_free(&config.blacklist);
 
     if (close(epollfd)) {
         ERR("close");